Use virtual_codebase for traversal #781
